maxIdleTime is long (delegate)

+++ integ-wget/src/main/java/il/co/topq/integframework/WgetClient.java
@@ -175,7 +175,7 @@ public class WgetClient implements Callable<String> {
 		return this;
 	}
 
-	public void setMaxIdleTime(int maxIdleTimeMillies) {
+	public void setMaxIdleTime(long maxIdleTimeMillies) {
 		module.getCliConnectionImpl().setMaxIdleTime(maxIdleTimeMillies);
 	}
